% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stan_model.R
\name{create_stan_model}
\alias{create_stan_model}
\title{create_stan_model}
\usage{
create_stan_model(formula, df_weights, family)
}
\arguments{
\item{formula, }{which specifies the model structure}

\item{df_weights}{is a data frame of weights (a vector of weights converted into a data frame)}

\item{family:}{distribution family of the model.}
}
\value{
model: the output is the compiled model.
}
\description{
This R function creates Stan models based on the specified formula, the data frame of weights, and family distribution.
}
\examples{
## Build models.
#choose some weights to test the model on:
weights <- c(2,1,0,4,2,2,2,1,1,3,8) #can be anything, just need the variable names

#add arbitrary labels (this is used just to build the model)
weight_labels <- as.factor(rep(c(0,1),length(weights))) 

#build and compile all base models (this saves time when running the models!)
normal_model_base <-create_stan_model('weights ~ 1', data.frame(weights=weights),"normal")
skewnormal_model_base <- create_stan_model('weights ~ 1', data.frame(weights=weights), "skew_normal")
poisson_model_base <- suppressMessages(create_stan_model('weights ~ 1', data.frame(weights=weights), "poisson"))
poisson_model_label <- suppressMessages(create_stan_model('weights ~ 1 + label', data.frame(weights=weights, label=weight_labels), "poisson"))

}
