% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stan_model.R
\name{fit_stan_model}
\alias{fit_stan_model}
\title{fit_stan_model}
\usage{
fit_stan_model(model, formula, df_weights, family)
}
\arguments{
\item{model}{which corresponds to the base model (normal_model_base, skewnormal_model_base, poisson_model_base, poisson_model_label)}

\item{formula}{specifies the model structure.}

\item{df_weights}{is a vector of weights}

\item{family}{is the distribution family of the model}
}
\value{
fit: the output is the fitted brms model.
}
\description{
This function fits the a Stan model to the given data, using the rstan package
}
\examples{
## Build models.
#choose some weights to test the model on:
weights <- c(2,1,0,4,2,2,2,1,1,3,8) #can be anything, just need the variable names

#add arbitrary labels (this is used just to build the model)
weight_labels <- as.factor(rep(c(0,1),length(weights))) 

#build and compile all base models (this saves time when running the models!)
normal_model_base <-create_stan_model('weights ~ 1', data.frame(weights=weights),"normal")

model.output<-fit_stan_model(normal_model_base,'weights ~ 1', data.frame(weights=weights), family="normal")
}
